requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==7.1
django-shopping-cart==0.1